requirements.txt
flask
gunicorn
numpy
pillow
pilmoji
prometheus_client
requests
